bhyve: allow overriding snapshot socket directory
By default bhyve(8) creates a snapshot socket in
"/var/run/bhyve/" (BHYVE_RUN_DIR).
As this is a system directory not writable by users,
this does not work when bhyve(8) is being started
as a non-root user.

Address that by allowing to override this directory.
In bhyve(8) it is done by setting 'rundir' option
with '-o rundir=<path>'.

In bhyvectl(8) it is done with '--rundir=<path>'.

MFC after:		1 month
Reviewed by:		bcr (manpages), bnovkov
Sponsored by:		The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D57494

+++ usr.sbin/bhyve/snapshot.c
@@ -1399,13 +1399,13 @@ IPC_COMMAND(ipc_cmd_set, checkpoint, vm_do_checkpoint);
  * Create the listening socket for IPC with bhyvectl
  */
 int
-init_checkpoint_thread(struct vmctx *ctx)
+init_checkpoint_thread(struct vmctx *ctx, const char *bhyve_run_dir)
 {
 	struct checkpoint_thread_info *checkpoint_info = NULL;
 	struct sockaddr_un addr;
 	int socket_fd;
 	pthread_t checkpoint_pthread;
-	int err;
+	int err, ret;
 #ifndef WITHOUT_CAPSICUM
 	cap_rights_t rights;
 #endif
@@ -1421,8 +1421,14 @@ init_checkpoint_thread(struct vmctx *ctx)
 
 	addr.sun_family = AF_UNIX;
 
-	snprintf(addr.sun_path, sizeof(addr.sun_path), "%s%s",
-		 BHYVE_RUN_DIR, vm_get_name(ctx));
+	ret = snprintf(addr.sun_path, sizeof(addr.sun_path), "%s/%s",
+	    bhyve_run_dir, vm_get_name(ctx));
+	if ((ret < 0) || ((size_t)ret >= sizeof(addr.sun_path))) {
+		EPRINTLN("%s: error setting socket path (%d)", __func__, ret);
+		err = -1;
+		goto fail;
+	}
+
 	addr.sun_len = SUN_LEN(&addr);
 	unlink(addr.sun_path);
